Turned build to a multi-project build.
Added a 'core' project which contains common classes and the central
API. More projects are going to be added for concrete file system
implementations.

Interestingly, after the move to a sub folder, WireMock was no longer
able to find resource files to be served. For this problem a
work-around has been added.