Changed return value of functions to create/update files and folders.
It is more generic to let the create functions return the IDs of newly
created elements rather than the elements themselves. Analogously, the
update functions just return Unit. Then the application can decide
whether it needs to resolve the new/updated elements or not.

Also fixed the missing implicit actor system parameter in the
updateFileContent() function.