Improvements of WireMockSupport trait.
Added a convenience function to define JSON responses. This function
automatically sets the correct Content-Type header for the response
entity. This is needed when the response is evaluated using Akka
HTTP's spray-json support.

Reworked the trait to start the mock server once at the beginning and
shut it down after all tests. It turned out that the server was not
fully reset, so that stub mappings from previous tests were still
active. There was an incorrect test case in DavFileSystemITSpec, which
was not detected due to this issue.