Added "writable" data classes to OneDriveJsonProtocol.
To generate JSON from model objects (which is required for some
requests), the existing classes from the protocol do not work well
because of most of their properties are irrelevant (because read-only)
for the serialization. They must, however, not be null because then
spray-json rejects the serialization. To work around this problem, a
simplified variant of DriveItem was introduced that contains only
relevant fields as Options.