Implemented file uploads in OneDriveFileSystem.
The functions to create new files and update the content of existing
files have now been implemented with the help of the new
OneDriveUpload module.

Some changes on other components were necessary: The upload chunk size
in the OneDriveConfig is now specified in bytes; the sizes of files
must be specified when creating new files; new model classes have been
added to OneDriveJsonProtocol.