OneDriveFileSystem offers a function to create an HTTP actor.
As the HttpSenderActor to be used by OneDriveFileSystem must fulfill
specific requirements, the companion object now offers a function to
create such an actor. The resulting actor supports multiple hosts (for
the API host and the up- and download servers), provided OAuth token
authentication against the API server, but not for others.