Added functionality to UriEncodingHelper related to URI components.
It is now possible to split a URI into components and decode them, and
to create a URI from components that need to be encoded first. This is
useful when dealing with URIs that refer to elements in a file system.
By splitting such URIs, a sequence of folder and file names can be
obtained; but the element names may contain characters not allowed in
URIs, so encoding is necessary.