Introducing the concept of a Spawner.
The Spawner trait allows creating typed actor instances. It is an
abstraction over the different ways such actors can be created, e.g.
from a typed or untyped actor context. The trait is going to be used
when the library creates actors on behalf of a client. The concrete
creation mechanism used by the client is then irrelevant for the
library.