Added a trait for creating configured request sender actors.
The idea is to have an HttpRequestSenderConfig class that defines the
properties of the new sender actor. This is passed to the factory,
which creates the new actor instance and decorates it with the
extensions desired.

This simplifies the creation of initialized request sender actors and
also makes it easier to test for client code.