Fixed a problem with discarding entities in RetryAfterExtension.
When failed requests are retried, and the original DiscardEntityMode
is Never, the response entities have to be discarded. (In all other
cases, this happens automatically.)

Unfortunately, the test class had to be rewritten because
BehaviorTestkit cannot be used when the actor under test requires a
classic actor system. (This is needed here to run the stream that
discards the bytes of the response entities.)