Added LRUCache class.
This is a map with a defined capacity that keeps the last recent used
entries. It is going to be used by the crypt module to cache paths in
a file system with encrypted names; but as it is of general use, it is
added to a new utils package in the core module.