HttpRequestSender now supports a ProxySelectorFunc argument.
Via this parameter, it is now possible to configure a proxy that is
used when sending requests.

An integration test has been added for this new functionality. This
was not trivial; here a real proxy server is used, and it is checked
whether it is invoked when a request is sent.