LocalFsModel: Aligned the functions to create new files or folders.
The functions now also accept an optional path parameter, so that the
folder or file path can be set. This can be used when creating objects
for update operations and makes the updateFolder() and updateFile()
functions obsolete.

These functions are now aligned to the corresponding ones provided by
other model implementations.