Changed description property from String to Option[String].
The description of files and folders is used rather infrequently; by
some file systems it is not even supported. So, instead of dealing
with null values, using the type Option[String] for this property
seems to be more appropriate. Actually, this is also the native type
for some file systems.