Fixed a deserialization problem in GoogleDriveFileSystem.
When creating a new file or folder, the response from the server
contains only a limited set of properties of the new element;
therefore, the deserialization to a full GoogleDriveJsonProtocol.File
object fails. To prevent this, now a FileReference is read - this is
sufficient, as only the ID of the new element is relevant.