OneDrive: Added special handling for files with size 0.
The standard upload logic did not work for such files (since the
Content-range header was rejected). Therefore, the upload of those
special files is now done via the API for small uploads with an empty
data source.

Note: There does not seem to be a possibility to create files without
content; so an upload is always required.