DavFileSystem now handles multi-status responses for PROPPATCH.
The PROPPATCH requests sent when creating or updating files and
folders can return 207 multi-status responses. So far, this status has
been treated as success. Now the entities of such responses are parsed
and checked whether they contain error codes for the single
operations. If this is the case, the whole operation must be
considered failed.