CryptService: Changed signature of decryptTextFromBase64().
As the decoding from Base64 can fail, the function now returns a Try.
The same is true for the decodeBase64() function used under the hood.

Handling this failure directly in CryptService is preferable than
doing this in client code; at the end of the day, all clients need to
handle potential exceptions.