Model: Added a new mapping function to FolderContent.
The new function allows parallel mapping of files and folders with a
function that can fail. The resulting FolderContent object only
contains the elements that were mapped successfully; a list with all
failures is returned as well.

With this function, it is possible to filter elements based on some
criteria. It is going to be used by CryptNamesFileSystem to deal with
incorrectly encoded elements in the file system.