DavParser: Started rework of XML parsing.
The goal of this refactoring is to replace the Scala XML module by
Java board means for XML parsing, namely SAX parsing. This should not
be significantly more complex, is likely to be more efficient, and can
be done without dependencies. The Scala XML module also seems to be
problematic when it comes to Scala 3 migration.

In this first step, the parsing of a folder and its elements was
rewritten to use a SAX parser.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>