onedrive: Dropped Authorization header for downloads/uploads.
Requests to download or upload files via special URLs must not have an
`Authorization` header. In the past, such a header was ignored. But
now, the OneDrive server responds with error code 401 when it is
present. Therefore, `OneDriveFileSystem` explicitly sets an empty
`Authorization` header, so that it is dropped by the Auth extension.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>