core: Reworked walk functionality to be more parallel.
The resolving of folders is now done in parallel to the iteration. If
the iteration comes to a point where it needs a folder which has not
yet been resolved, it returns a special waiting state. The
`WalkSource` implementation then calls the walk function only after
further folder results are available.