core: Introduced a config class for walk operations.
The functions to create walk sources already expect a larger number of
parameters. This may be increased in the future if more features or
configuration options are added. Therefore, it makes sense to have a
configuration class that collects all these parameters and is able to
provide meaningful defaults for optional parameters.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>