core: Introduced `WalkFolderState` helper class.
This class holds information about already resolved folders during a
walk operation. It allows that the state about folders is managed by
the walk source and simplifies state handling in the walk functions.

Because the Scala 3 compiler was complaining about incompatible types,
the types related to walk functions and results have been moved to
members of the `WalkSource` class. This allows further
simplifications, as the complex type parameters are no longer needed.
(They were also the reason for the problems of the Scala 3 compiler.)

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>