core: Introduced properties to tweak folder resolving in walks.
It is now possible to define how many folders should be resolved in
parallel, and how many should be fetched ahead. Especially for walk
operations in BFS order, this can speed up iterations.

In the original implementation, these parameters had been hard-coded.
It turned out that for BFS traversals with many subfolders, a very
high load was put on servers.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>